Changes time format from datetime to timestamps
The tracks used datetime objects as timestamps because of legacy
reasons. As we are no longer using or supporting the legacy database
system we can now remove the overhead with different datetime objects
(python, numpy and pandas) and use the timestamps from *bb_binary*
directly. Use `bb_binary.to_datetime()` to convert these timestamps
to a human readable format.